@@ -16,38 +16,38 @@
   ],
   "license": "MIT",
   "engines": {
-    "node": ">=18"
+    "node": ">=20"
   },
   "dependencies": {
-    "@mozaic-ds/gatsby-theme-styleguide": "^1.63.0",
-    "adm-zip": "^0.5.10",
-    "chromedriver": "^120.0.1",
-    "core-js": "^3.34.0",
-    "gatsby": "^5.13.1",
+    "@mozaic-ds/gatsby-theme-styleguide": "^1.70.1",
+    "adm-zip": "^0.5.14",
+    "chromedriver": "^126.0.1",
+    "core-js": "^3.37.1",
+    "gatsby": "^5.13.6",
     "gatsby-plugin-canonical-urls": "^5.13.1",
-    "gatsby-plugin-sitemap": "^6.13.0",
+    "gatsby-plugin-sitemap": "^6.13.1",
     "glob": "^8.0.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "request-promise": "^4.2.6",
     "rimraf": "^3.0.2",
     "svgo": "^1.3.2",
     "xstate": "^4.38.3"
   },
   "devDependencies": {
-    "@commitlint/cli": "^18.4.3",
-    "@commitlint/config-conventional": "^18.4.3",
+    "@commitlint/cli": "^19.3.0",
+    "@commitlint/config-conventional": "^19.2.2",
     "all-contributors-cli": "^6.26.1",
     "cp-file": "^11.0.0",
-    "husky": "^8.0.3",
+    "husky": "^9.0.11",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-cli": "^29.7.0",
     "lerna": "^7.4.2",
-    "lint-staged": "^15.2.0",
+    "lint-staged": "^15.2.7",
     "mock-fs": "^5.2.0",
-    "plop": "^3.1.2",
-    "prettier": "^3.1.1",
+    "plop": "^4.0.1",
+    "prettier": "^3.3.2",
     "svg-parser": "^2.0.4",
     "yarn-audit-fix": "^10.0.7"
   },
@@ -74,7 +74,7 @@
     "unit": "jest --projects ./packages/*/jest.config.js --coverage --verbose --color --no-cache",
     "unit:watch": "jest --watch --silent",
     "unit:coverage": "jest --coverage --collectCoverageFrom='**/*.js'",
-    "prepare": "husky install"
+    "prepare": "husky || true"
   },
   "packageManager": "yarn@1.22.19"
 }
