@@ -1,37 +1,37 @@
 {
   "name": "@mozaic-ds/gatsby-theme-styleguide",
-  "version": "1.70.1",
+  "version": "1.75.0",
   "main": "index.js",
   "author": "gael-boyenval (gaelboyenval@icloud.com)",
   "license": "MIT",
   "dependencies": {
     "@code-hike/mdx": "^0.9.0",
-    "@mdx-js/react": "^3.0.0",
-    "@mozaic-ds/gatsby-github-release": "^1.70.0",
+    "@mdx-js/react": "^3.0.1",
+    "@mozaic-ds/gatsby-github-release": "^1.75.0",
     "gatsby-plugin-flexsearch": "^1.0.3",
-    "gatsby-plugin-mdx": "^5.13.0",
-    "gatsby-plugin-sass": "^6.13.0",
-    "gatsby-plugin-sharp": "^5.13.0",
-    "gatsby-remark-autolink-headers": "^6.13.0",
-    "gatsby-remark-images": "^7.13.0",
+    "gatsby-plugin-mdx": "^5.13.1",
+    "gatsby-plugin-sass": "^6.13.1",
+    "gatsby-plugin-sharp": "^5.13.1",
+    "gatsby-remark-autolink-headers": "^6.13.1",
+    "gatsby-remark-images": "^7.13.1",
     "gatsby-remark-static-images": "^1.2.1",
-    "gatsby-source-filesystem": "^5.13.0",
-    "gatsby-transformer-json": "^5.13.0",
+    "gatsby-source-filesystem": "^5.13.1",
+    "gatsby-transformer-json": "^5.13.1",
     "remark-gfm": "^3.0.1",
-    "sass": "^1.69.5"
+    "sass": "^1.77.6"
   },
   "peerDependencies": {
-    "gatsby": "^5.13.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "gatsby": "^5.13.6",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "@sindresorhus/slugify": "^2.2.1",
-    "@testing-library/jest-dom": "^6.1.5",
-    "core-js": "^3.34.0",
+    "@testing-library/jest-dom": "^6.4.6",
+    "core-js": "^3.37.1",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.7",
+    "enzyme-adapter-react-16": "^1.15.8",
     "jest": "^29.7.0",
-    "react-test-renderer": "^18.2.0"
+    "react-test-renderer": "^18.3.1"
   }
 }
