@@ -8,7 +8,7 @@ It was auto-generated by "iconCompiler" script
 
 function generatePaths(arr, iconName) {
   const objectToReturn = arr.map(
-    (elem) => `{${generateTagObject(elem, iconName)}}`
+    (elem) => `{${generateTagObject(elem, iconName)}}`,
   )
   return `[${objectToReturn}]`
 }
@@ -17,26 +17,32 @@ function generateTagObject(obj, iconName) {
   let objString = ''
   let tagObj = {}
   const filename = iconName.toLowerCase()
+  const iconBaseId = `${filename}-id`
   const tagName = obj.tagName
   const properties = obj.properties
+  const propKeys = Object.keys(properties)
   const children = obj.children
-  const hasProperties = Object.keys(properties).length !== 0
+  const hasProperties = propKeys.length !== 0
   const hasChildren = children.length > 0
 
   if (tagName) tagObj.tagName = tagName
 
   if (hasProperties) {
-    const customFillId = `fill-${filename}`
+    if (properties['clip-path'] || properties.fill || properties.filter) {
+      const propValue =
+        properties['clip-path'] || properties.fill || properties.filter
 
-    if (
-      (tagName === 'path' || tagName === 'rect') &&
-      properties.fill &&
-      properties.fill.includes('url')
-    ) {
-      properties.fill = `url(#${customFillId})`
+      if (propValue.includes('url')) {
+        const id = propValue.split('#')[1].slice(0, -1)
+
+        const propKey = propKeys.find((key) => properties[key] === propValue)
+
+        properties[propKey] = `url(#${iconBaseId}-${id})`
+      }
     }
-    if (properties.id && tagName === 'linearGradient') {
-      properties.id = customFillId
+
+    if (properties.id) {
+      properties.id = `${iconBaseId}-${properties.id}`
     }
 
     tagObj.attrs = JSON.stringify(properties)
@@ -65,7 +71,7 @@ const iconModule = ({ fileName, data }) => {
 
   if (!sizeRem || !content || content.length === 0) {
     console.error(
-      `✗ ERROR : there was a problem when transforming the ${fileName} file into a JS module`
+      `✗ ERROR : there was a problem when transforming the ${fileName} file into a JS module`,
     )
 
     return null
