@@ -7,4 +7,4 @@
   export let size = "1.5rem";
 </script>
 
-<svg aria-hidden="true" width={size} height={size} id={id}  style={style} class={className} fill={fill} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 60 24"><path d="M19.63 11.065c-1.906 0-3.458-1.58-3.458-3.594 0-2.013 1.552-3.597 3.459-3.597 1.893 0 3.455 1.584 3.455 3.597 0 2.014-1.561 3.594-3.455 3.594zm0-9.06c-3 0-5.35 2.4-5.35 5.466 0 3.065 2.35 5.466 5.35 5.466 2.994 0 5.344-2.401 5.344-5.466-.002-3.068-2.352-5.465-5.343-5.465zm27.519 5.321c-.018-2.926-2.34-5.308-5.18-5.308-2.922 0-5.214 2.39-5.214 5.44a5.59 5.59 0 001.508 3.83c.986 1.051 2.287 1.63 3.672 1.63 1.58 0 3.793-.773 4.847-2.965l.027-.07h-2.12l-.013.016c-.569.692-1.642 1.16-2.664 1.16-1.3 0-2.857-.872-3.284-2.84h8.38c.02-.204.04-.612.04-.893zm-8.36-.846c.258-1.32 1.498-2.608 3.165-2.608 1.665 0 2.903 1.288 3.159 2.608h-6.324zm-7.911-2.669c-1.888 0-2.85 1.032-2.85 3.07v6.04h-1.94V6.739c.02-1.084.208-1.85.643-2.57C27.244 3.361 28.465 2 30.878 2c2.407 0 3.63 1.36 4.143 2.168.43.72.625 1.486.636 2.57v6.184H33.72v-6.04c0-2.041-.953-3.07-2.841-3.07zM57.23 2.006l-5.653 14.056H49.56l1.565-3.9-4.178-10.156h2.13l3.059 7.636L55.2 2.006h2.03zM7.847 18.145c0 .438-.099.826-.296 1.168a1.9 1.9 0 01-.812.772L7.78 22H6.222l-.93-1.673h-.718V22h-1.37v-6.037h2.52c.615 0 1.121.206 1.523.617.398.41.6.931.6 1.565zm-3.273-1.044v2.087h.938a.879.879 0 00.681-.293c.179-.196.266-.446.266-.75s-.089-.555-.266-.75a.876.876 0 00-.681-.293h-.938V17.1zM12.387 22l-.203-.732h-2.096L9.885 22H8.439l2.004-6.037h1.378L13.834 22h-1.447zm-1.995-1.829h1.489l-.744-2.682-.745 2.682zM15.212 22v-4.794h-1.624v-1.243h4.617v1.243h-1.631V22h-1.362zm7.468-6.037h1.513L21.26 22h-1.523l1.1-2.277-2.098-3.76h1.59l1.26 2.381 1.092-2.38zM7.956 12.908c2.949 0 5.34-2.438 5.34-5.445s-2.391-5.445-5.34-5.445c-2.95 0-5.34 2.438-5.34 5.445s2.39 5.445 5.34 5.445z"/><path d="M10.218 6.095h-.904v-.49c0-.637-.508-1.155-1.132-1.155h-.67c-.624 0-1.131.518-1.131 1.155v.49h-.905a.233.233 0 00-.23.236c0 .13.102.235.23.235h4.744a.03.03 0 01.021.01l-.755 3.26a.25.25 0 00-.005.083.03.03 0 01-.007.025.03.03 0 01-.025.011H6.271a.233.233 0 00-.23.236c0 .13.103.235.23.235H9.45a.487.487 0 00.37-.17.51.51 0 00.124-.346l.761-3.29a.25.25 0 00.005-.084.498.498 0 00-.491-.44zm-3.376-.49c0-.377.301-.684.67-.684h.671c.37 0 .67.307.67.684v.49h-2.01v-.49z"/><path d="M6.617 9.248a.184.184 0 00-.181-.185H5.4c-.1 0-.182.083-.182.185s.082.185.182.185h1.035c.1 0 .181-.083.181-.185zM5.36 7.521h1.036c.1 0 .181-.083.181-.185a.184.184 0 00-.181-.185H5.36c-.1 0-.182.083-.182.185s.082.185.182.185zm1.648.783a.184.184 0 00-.181-.185H4.526c-.1 0-.182.083-.182.185s.082.185.182.185h2.3a.182.182 0 00.182-.185z"/></svg>
+<svg aria-hidden="true" width={size} height={size} id={id}  style={style} class={className} fill={fill} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 60 24"><path d="M28.028 6.881c0-2.038.962-3.07 2.85-3.07 1.888 0 2.841 1.03 2.841 3.07v6.04h1.938V6.739c-.011-1.084-.206-1.85-.636-2.57C34.507 3.361 33.285 2 30.878 2c-2.413 0-3.634 1.36-4.147 2.168-.435.72-.623 1.486-.643 2.57v6.184h1.94v-6.04z"/><path fill-rule="evenodd" d="M19.63 2.006c-3 0-5.35 2.4-5.35 5.465s2.35 5.466 5.35 5.466c2.994 0 5.343-2.401 5.343-5.466-.002-3.068-2.351-5.465-5.342-5.465zm0 9.059c-1.907 0-3.458-1.58-3.458-3.594 0-2.013 1.551-3.597 3.459-3.597 1.893 0 3.454 1.584 3.454 3.597 0 2.014-1.56 3.594-3.454 3.594zm27.518-3.739c-.017-2.926-2.34-5.308-5.179-5.308-2.922 0-5.214 2.39-5.214 5.44 0 1.433.538 2.791 1.507 3.83.986 1.051 2.288 1.63 3.672 1.63 1.58 0 3.794-.773 4.847-2.965l.027-.07H44.69l-.014.016c-.57.692-1.642 1.16-2.665 1.16-1.3 0-2.857-.872-3.284-2.84h8.381c.02-.204.04-.612.04-.893zM38.79 6.48c.258-1.32 1.498-2.608 3.165-2.608 1.665 0 2.903 1.288 3.159 2.608H38.79z"/><path d="M57.23 2.006l-5.654 14.056H49.56l1.566-3.9-4.178-10.156h2.13l3.059 7.636L55.2 2.006h2.03z"/><path fill-rule="evenodd" d="M7.551 19.313c.197-.341.295-.73.295-1.168 0-.634-.201-1.154-.6-1.565-.401-.411-.907-.617-1.522-.617H3.203V22h1.37v-1.673h.719L6.222 22H7.78l-1.04-1.915c.344-.172.615-.43.812-.772zm-2.978-.125v-2.086h.939c.276 0 .504.097.68.293.178.195.267.446.267.75s-.088.554-.266.75a.879.879 0 01-.681.293h-.939zM12.387 22l-.203-.732h-2.097L9.884 22H8.44l2.004-6.037h1.378L13.834 22h-1.447zm-1.995-1.829h1.489l-.745-2.682-.744 2.682z"/><path d="M15.212 17.206V22h1.361v-4.794h1.632v-1.243h-4.617v1.243h1.624zm7.468-1.243h1.513L21.258 22h-1.522l1.1-2.277-2.099-3.76h1.59l1.26 2.381 1.092-2.38zM7.51 4.92a.679.679 0 00-.67.684v.491h2.012v-.49a.679.679 0 00-.671-.684h-.67z"/><path fill-rule="evenodd" d="M13.296 7.463c0 3.007-2.391 5.445-5.34 5.445-2.95 0-5.34-2.438-5.34-5.445s2.39-5.445 5.34-5.445c2.949 0 5.34 2.438 5.34 5.445zM9.31 6.095h.905c.249 0 .46.19.49.44a.25.25 0 01-.004.084L9.94 9.91a.51.51 0 01-.124.347.487.487 0 01-.37.17H6.27a.233.233 0 01-.23-.236c0-.13.102-.235.23-.235h3.178c.01 0 .018-.003.024-.011a.03.03 0 00.008-.025.25.25 0 01.005-.083l.755-3.26a.03.03 0 00-.022-.01H5.474a.233.233 0 01-.23-.236c0-.13.102-.235.23-.235h.904v-.49c0-.637.508-1.155 1.132-1.155h.67c.624 0 1.131.518 1.131 1.154v.491zM6.434 9.063c.1 0 .181.083.181.185a.184.184 0 01-.181.185H5.399a.184.184 0 01-.182-.185c0-.102.082-.185.182-.185h1.035zM5.359 7.52h1.035c.1 0 .181-.083.181-.185a.184.184 0 00-.181-.185H5.359c-.1 0-.182.083-.182.185s.082.185.182.185zm1.465.597c.1 0 .182.083.182.185a.182.182 0 01-.183.185h-2.3a.184.184 0 01-.181-.185c0-.102.081-.185.181-.185h2.301z"/></svg>
