@@ -7,4 +7,4 @@
   export let size = "2rem";
 </script>
 
-<svg aria-hidden="true" width={size} height={size} id={id}  style={style} class={className} fill={fill} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 82 32"><path d="M26.514 14.691c-2.67 0-4.842-2.212-4.842-5.03 0-2.82 2.172-5.037 4.842-5.037 2.651 0 4.837 2.217 4.837 5.036 0 2.819-2.186 5.031-4.837 5.031zm0-12.683c-4.202 0-7.491 3.36-7.491 7.652 0 4.29 3.29 7.652 7.49 7.652 4.192 0 7.48-3.362 7.48-7.652-.002-4.295-3.291-7.652-7.48-7.652zm38.525 7.449c-.025-4.098-3.276-7.432-7.251-7.432-4.092 0-7.3 3.345-7.3 7.616 0 2.006.754 3.908 2.11 5.363 1.381 1.471 3.203 2.28 5.141 2.28 2.213 0 5.31-1.08 6.786-4.15l.038-.098h-2.967l-.018.022c-.797.969-2.3 1.625-3.731 1.625-1.82 0-4-1.222-4.597-3.977h11.732a16.3 16.3 0 00.057-1.249zM53.336 8.271c.36-1.847 2.096-3.65 4.43-3.65 2.331 0 4.065 1.803 4.423 3.65h-8.853zM42.26 4.536c-2.643 0-3.99 1.444-3.99 4.298v8.456h-2.715V8.634c.027-1.518.29-2.591.899-3.598C37.172 3.905 38.882 2 42.26 2c3.37 0 5.082 1.905 5.8 3.036.603 1.007.875 2.08.891 3.598v8.656h-2.713V8.834c0-2.857-1.335-4.298-3.978-4.298zm36.894-2.528l-7.916 19.68h-2.824l2.191-5.46-5.848-14.22h2.982l4.282 10.69 4.29-10.69h2.843zM10.016 24.603c0 .613-.138 1.157-.414 1.636a2.668 2.668 0 01-1.136 1.08L9.92 30H7.742l-1.303-2.343H5.434V30H3.515v-8.451h3.53c.86 0 1.57.287 2.131.863.558.575.84 1.304.84 2.191zm-4.582-1.462v2.922h1.313c.388 0 .704-.136.954-.41.25-.275.373-.624.373-1.05 0-.426-.125-.777-.373-1.05a1.227 1.227 0 00-.954-.41H5.434v-.002zM16.372 30l-.284-1.025h-2.935L12.87 30h-2.024l2.806-8.451h1.93L18.398 30h-2.027zm-2.792-2.56h2.084l-1.042-3.755-1.042 3.755zM20.328 30v-6.712h-2.274v-1.74h6.464v1.74h-2.284V30h-1.906zm10.454-8.451h2.12L28.791 30H26.66l1.54-3.188-2.938-5.263h2.227l1.763 3.333 1.53-3.333zm-20.613-4.277c4.129 0 7.476-3.413 7.476-7.624 0-4.21-3.347-7.623-7.476-7.623-4.13 0-7.477 3.413-7.477 7.623 0 4.21 3.348 7.624 7.477 7.624z"/><path d="M13.336 7.734H12.07v-.688c0-.891-.71-1.616-1.585-1.616h-.937c-.874 0-1.585.725-1.585 1.616v.688H6.697a.326.326 0 00-.323.33c0 .181.145.328.323.328h6.641c.01 0 .022.005.03.014l-1.057 4.564a.351.351 0 00-.006.117.04.04 0 01-.011.035.044.044 0 01-.035.015H7.811a.326.326 0 00-.323.33c0 .182.144.329.323.329h4.448a.682.682 0 00.517-.237.715.715 0 00.175-.485l1.065-4.607a.35.35 0 00.007-.116.698.698 0 00-.687-.617zm-4.727-.688a.95.95 0 01.94-.957h.938a.95.95 0 01.939.957v.688H8.609v-.688z"/><path d="M8.295 12.147a.257.257 0 00-.254-.259h-1.45a.257.257 0 00-.253.26c0 .142.114.258.254.258H8.04c.14 0 .254-.116.254-.259zM6.536 9.73h1.449c.14 0 .254-.117.254-.26a.257.257 0 00-.254-.259h-1.45a.257.257 0 00-.253.26c0 .142.114.259.254.259zm2.306 1.095a.257.257 0 00-.254-.259H5.367a.257.257 0 00-.254.26c0 .142.114.258.254.258h3.219a.255.255 0 00.256-.259z"/></svg>
+<svg aria-hidden="true" width={size} height={size} id={id}  style={style} class={className} fill={fill} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 82 32"><path d="M38.27 8.834c0-2.854 1.347-4.298 3.99-4.298 2.643 0 3.978 1.441 3.978 4.298v8.456h2.713V8.634c-.016-1.518-.288-2.591-.89-3.598C47.34 3.905 45.63 2 42.26 2c-3.378 0-5.087 1.905-5.806 3.036-.608 1.007-.872 2.08-.9 3.598v8.656h2.717V8.834z"/><path fill-rule="evenodd" d="M26.514 2.008c-4.202 0-7.491 3.36-7.491 7.652 0 4.29 3.29 7.652 7.49 7.652 4.192 0 7.48-3.362 7.48-7.652-.002-4.295-3.291-7.652-7.48-7.652zm0 12.683c-2.67 0-4.842-2.212-4.842-5.03 0-2.82 2.172-5.037 4.842-5.037 2.651 0 4.837 2.217 4.837 5.036 0 2.819-2.186 5.031-4.837 5.031zm38.525-5.234c-.025-4.098-3.276-7.432-7.251-7.432-4.092 0-7.3 3.345-7.3 7.616 0 2.006.754 3.908 2.11 5.363 1.381 1.471 3.203 2.28 5.141 2.28 2.213 0 5.31-1.08 6.786-4.15l.038-.098h-2.967l-.018.022c-.797.969-2.3 1.625-3.731 1.625-1.82 0-4-1.222-4.597-3.977h11.732a16.3 16.3 0 00.057-1.249zM53.336 8.271c.36-1.847 2.096-3.65 4.43-3.65 2.331 0 4.065 1.803 4.423 3.65h-8.853z"/><path d="M79.154 2.008l-7.916 19.68h-2.824l2.191-5.46-5.848-14.22h2.982l4.282 10.69 4.29-10.69h2.843z"/><path fill-rule="evenodd" d="M9.602 26.239c.276-.479.414-1.023.414-1.636 0-.887-.282-1.616-.84-2.191-.562-.576-1.27-.863-2.132-.863H3.515V30h1.919v-2.343h1.005L7.742 30h2.18l-1.457-2.68a2.668 2.668 0 001.137-1.081zm-4.168-.176v-2.92h1.313c.388 0 .706.136.954.41.248.273.373.624.373 1.05 0 .426-.123.775-.373 1.05a1.23 1.23 0 01-.954.41H5.434zM16.372 30l-.284-1.025h-2.935L12.87 30h-2.024l2.806-8.451h1.93L18.398 30h-2.027zm-2.792-2.56h2.084l-1.042-3.755-1.042 3.755z"/><path d="M20.328 23.288V30h1.906v-6.712h2.284v-1.74h-6.464v1.74h2.274zm10.454-1.739h2.12L28.791 30H26.66l1.54-3.188-2.938-5.263h2.227l1.763 3.333 1.53-3.333zM9.546 6.089a.95.95 0 00-.94.957v.687h2.817v-.687a.95.95 0 00-.939-.957h-.938z"/><path fill-rule="evenodd" d="M17.645 9.648c0 4.21-3.347 7.624-7.476 7.624-4.13 0-7.477-3.413-7.477-7.624 0-4.21 3.348-7.623 7.477-7.623 4.129 0 7.476 3.413 7.476 7.623zm-5.578-1.915h1.266c.349 0 .644.266.687.617a.35.35 0 01-.006.117l-1.066 4.606a.714.714 0 01-.175.486.682.682 0 01-.516.237h-4.45a.326.326 0 01-.322-.33c0-.182.144-.33.323-.33h4.449c.015 0 .025-.003.034-.014a.04.04 0 00.01-.036.35.35 0 01.007-.116l1.058-4.565a.042.042 0 00-.03-.013H6.694a.326.326 0 01-.323-.33c0-.182.144-.329.323-.329H7.96v-.687c0-.891.71-1.616 1.585-1.616h.936c.874 0 1.585.725 1.585 1.616v.687zm-4.029 4.155c.14 0 .254.116.254.259a.257.257 0 01-.254.259H6.59a.257.257 0 01-.254-.26c0-.142.114-.258.254-.258h1.45zM6.533 9.729h1.45c.14 0 .253-.116.253-.259a.257.257 0 00-.254-.259H6.533a.257.257 0 00-.254.26c0 .142.114.258.254.258zm2.052.837c.14 0 .254.116.254.259a.255.255 0 01-.256.259h-3.22a.257.257 0 01-.253-.26c0-.142.114-.258.254-.258h3.221z"/></svg>
