@@ -7,4 +7,4 @@
   export let size = "3rem";
 </script>
 
-<svg aria-hidden="true" width={size} height={size} id={id}  style={style} class={className} fill={fill} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 125 48"><path d="M40.28 21.943c-4.196 0-7.61-3.476-7.61-7.906s3.414-7.914 7.61-7.914c4.166 0 7.6 3.485 7.6 7.914 0 4.43-3.434 7.906-7.6 7.906zm0-19.93c-6.603 0-11.772 5.279-11.772 12.024 0 6.741 5.17 12.025 11.772 12.025 6.585 0 11.754-5.284 11.754-12.025-.004-6.75-5.173-12.024-11.754-12.024zm60.539 11.705c-.038-6.439-5.147-11.68-11.394-11.68-6.43 0-11.471 5.258-11.471 11.97 0 3.152 1.184 6.14 3.316 8.427 2.17 2.312 5.033 3.584 8.079 3.584 3.476 0 8.344-1.7 10.663-6.521l.059-.156H95.41l-.03.035c-1.252 1.522-3.612 2.553-5.862 2.553-2.86 0-6.286-1.92-7.225-6.25h18.437a25.82 25.82 0 00.089-1.962zm-18.39-1.863c.566-2.903 3.294-5.736 6.962-5.736 3.663 0 6.387 2.833 6.95 5.736H82.428zm-17.406-5.87c-4.153 0-6.268 2.269-6.268 6.754v13.288h-4.268V12.425c.043-2.385.457-4.071 1.413-5.654C57.03 4.993 59.715 2 65.023 2c5.296 0 7.986 2.993 9.115 4.77.948 1.583 1.375 3.27 1.4 5.654v13.604h-4.263V12.739c0-4.49-2.098-6.754-6.252-6.754zM123 2.013l-12.439 30.924h-4.437l3.443-8.579-9.191-22.345h4.686l6.729 16.799 6.742-16.8H123zM14.355 37.519c0 .963-.217 1.818-.65 2.57a4.192 4.192 0 01-1.787 1.698L14.206 46H10.78l-2.047-3.681h-1.58V46H4.139V32.72h5.546c1.354 0 2.467.452 3.35 1.355.876.904 1.32 2.05 1.32 3.444zm-7.201-2.298v4.593h2.064c.61 0 1.107-.215 1.499-.646.393-.431.585-.98.585-1.649 0-.67-.196-1.221-.585-1.65-.39-.43-.89-.645-1.499-.645H7.154v-.003zM24.344 46l-.447-1.611h-4.612L18.838 46h-3.18l4.409-13.28h3.031L27.528 46h-3.184zm-4.39-4.023h3.276l-1.637-5.9-1.638 5.9zM30.56 46V35.452h-3.574V32.72h10.158v2.733h-3.59V46h-2.995zm16.427-13.28h3.33L43.861 46h-3.35l2.42-5.01-4.616-8.27h3.499l2.771 5.237 2.402-5.238zm-32.392-6.722c6.488 0 11.748-5.363 11.748-11.98 0-6.616-5.26-11.979-11.748-11.979-6.489 0-11.749 5.363-11.749 11.98 0 6.616 5.26 11.98 11.749 11.98z"/><path d="M19.572 11.01h-1.99V9.93c0-1.4-1.117-2.54-2.49-2.54H13.62c-1.374 0-2.49 1.14-2.49 2.54v1.08H9.14a.512.512 0 00-.508.518c0 .286.227.517.508.517h10.435a.07.07 0 01.048.02L17.96 19.24a.55.55 0 00-.01.183.064.064 0 01-.017.055.063.063 0 01-.054.024h-6.99a.512.512 0 00-.508.518c0 .286.226.518.507.518h6.99c.312 0 .607-.135.813-.373.183-.21.28-.483.274-.763l1.675-7.239a.55.55 0 00.01-.182 1.096 1.096 0 00-1.08-.97zm-7.428-1.08c0-.828.664-1.504 1.476-1.504h1.475c.812 0 1.475.676 1.475 1.504v1.08h-4.426V9.93z"/><path d="M11.65 17.945a.404.404 0 00-.399-.407H8.974c-.22 0-.4.183-.4.407 0 .225.18.408.4.408h2.277c.22 0 .4-.183.4-.408zm-2.764-3.798h2.277c.22 0 .4-.183.4-.408a.404.404 0 00-.4-.407H8.886c-.22 0-.4.183-.4.407 0 .225.18.408.4.408zm3.624 1.721a.404.404 0 00-.4-.407H7.05c-.22 0-.4.183-.4.407 0 .224.18.407.4.407h5.058a.4.4 0 00.403-.407z"/></svg>
+<svg aria-hidden="true" width={size} height={size} id={id}  style={style} class={className} fill={fill} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 125 48"><path d="M58.755 12.74c0-4.486 2.114-6.755 6.268-6.755s6.251 2.264 6.251 6.754v13.288h4.264V12.425c-.025-2.385-.453-4.071-1.4-5.654C73.008 4.993 70.318 2 65.023 2c-5.308 0-7.994 2.993-9.123 4.77-.956 1.583-1.37 3.27-1.413 5.654v13.604h4.267V12.739z"/><path fill-rule="evenodd" d="M40.28 2.013c-6.603 0-11.772 5.279-11.772 12.024 0 6.741 5.169 12.025 11.771 12.025 6.586 0 11.755-5.284 11.755-12.025-.005-6.75-5.173-12.024-11.755-12.024zm0 19.93c-4.197 0-7.61-3.476-7.61-7.906s3.413-7.914 7.61-7.914c4.166 0 7.6 3.485 7.6 7.914 0 4.43-3.434 7.906-7.6 7.906zm60.539-8.225c-.039-6.439-5.148-11.68-11.395-11.68-6.43 0-11.47 5.258-11.47 11.97 0 3.152 1.183 6.14 3.315 8.427 2.17 2.312 5.033 3.584 8.08 3.584 3.476 0 8.344-1.7 10.662-6.521l.06-.156h-4.662l-.03.035c-1.252 1.522-3.612 2.553-5.862 2.553-2.86 0-6.285-1.92-7.224-6.25h18.437a25.82 25.82 0 00.089-1.962zm-18.39-1.863c.566-2.903 3.294-5.736 6.961-5.736 3.663 0 6.387 2.833 6.95 5.736H82.428z"/><path d="M123 2.013l-12.44 30.924h-4.437l3.443-8.579-9.191-22.345h4.686l6.73 16.799 6.742-16.8H123z"/><path fill-rule="evenodd" d="M13.704 40.09c.433-.753.65-1.608.65-2.571 0-1.394-.443-2.54-1.32-3.444-.883-.904-1.996-1.356-3.35-1.356H4.138V46h3.015v-3.681h1.58L10.782 46h3.424l-2.287-4.213a4.191 4.191 0 001.786-1.697zm-6.55-.277v-4.589h2.064c.609 0 1.11.214 1.498.646.39.428.586.98.586 1.649 0 .67-.193 1.218-.585 1.65-.393.43-.89.645-1.5.645H7.154zM24.343 46l-.447-1.611h-4.612L18.838 46h-3.181l4.409-13.28h3.032L27.528 46h-3.185zm-4.389-4.023h3.276l-1.638-5.9-1.638 5.9z"/><path d="M30.559 35.452V46h2.994V35.452h3.59V32.72H26.986v2.733h3.574zm16.428-2.732h3.33L43.86 46h-3.35l2.42-5.01-4.616-8.27h3.5l2.77 5.237 2.403-5.238zM13.615 8.425c-.812 0-1.475.676-1.475 1.505v1.08h4.426V9.93c0-.829-.664-1.505-1.476-1.505h-1.475z"/><path fill-rule="evenodd" d="M26.343 14.019c0 6.616-5.26 11.98-11.749 11.98-6.488 0-11.748-5.364-11.748-11.98 0-6.617 5.26-11.98 11.748-11.98 6.489 0 11.749 5.363 11.749 11.98zm-8.766-3.01h1.99c.548 0 1.012.418 1.08.97a.548.548 0 01-.01.183l-1.676 7.239c.007.28-.091.552-.274.762a1.07 1.07 0 01-.812.373h-6.99a.512.512 0 01-.508-.518c0-.286.226-.517.507-.517h6.991c.024 0 .04-.007.054-.024a.064.064 0 00.017-.055.552.552 0 01.01-.183l1.662-7.174a.066.066 0 00-.048-.02H9.135a.512.512 0 01-.508-.518c0-.286.227-.518.508-.518h1.99V9.93c0-1.4 1.116-2.539 2.49-2.539h1.472c1.374 0 2.49 1.139 2.49 2.54v1.08zm-6.33 6.529c.22 0 .399.183.399.407 0 .224-.18.407-.4.407H8.97c-.22 0-.4-.183-.4-.407 0-.224.18-.407.4-.407h2.277zM8.88 14.146h2.277c.22 0 .4-.183.4-.407a.404.404 0 00-.4-.407H8.881c-.22 0-.4.183-.4.407 0 .224.18.407.4.407zm3.225 1.314c.22 0 .4.183.4.407a.4.4 0 01-.404.408H7.044c-.22 0-.4-.183-.4-.407 0-.225.18-.408.4-.408h5.062z"/></svg>
