@@ -7,4 +7,4 @@
   export let size = "4rem";
 </script>
 
-<svg aria-hidden="true" width={size} height={size} id={id}  style={style} class={className} fill={fill} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 168 64"><path d="M53.046 29.195c-5.722 0-10.377-4.74-10.377-10.78 0-6.04 4.655-10.793 10.377-10.793 5.681 0 10.364 4.752 10.364 10.793 0 6.04-4.683 10.78-10.364 10.78zm0-27.178c-9.004 0-16.052 7.2-16.052 16.398 0 9.192 7.048 16.397 16.052 16.397 8.98 0 16.028-7.205 16.028-16.397C69.07 9.21 62.02 2.017 53.046 2.017zm82.553 15.963c-.052-8.781-7.019-15.927-15.537-15.927-8.768 0-15.643 7.17-15.643 16.32 0 4.3 1.615 8.375 4.522 11.493 2.959 3.152 6.864 4.887 11.017 4.887 4.741 0 11.379-2.317 14.541-8.893l.08-.211h-6.356l-.041.047c-1.707 2.076-4.925 3.481-7.993 3.481-3.9 0-8.572-2.617-9.852-8.521h25.141a34.9 34.9 0 00.121-2.676zm-25.078-2.541c.773-3.958 4.494-7.822 9.494-7.822 4.995 0 8.71 3.864 9.477 7.822h-18.971zM86.787 7.434c-5.664 0-8.548 3.094-8.548 9.21v18.12h-5.82V16.216c.058-3.252.624-5.552 1.927-7.71C75.886 6.082 79.55 2 86.787 2c7.221 0 10.89 4.082 12.43 6.505 1.292 2.158 1.874 4.458 1.909 7.71v18.55h-5.814v-18.12c0-6.123-2.86-9.21-8.525-9.21zm79.059-5.416l-16.962 42.169h-6.051l4.695-11.698-12.533-30.471h6.39l9.176 22.907 9.194-22.907h6.091zM17.693 50.435c0 1.313-.295 2.48-.886 3.505a5.716 5.716 0 01-2.436 2.315L17.491 62h-4.67l-2.792-5.02H7.874V62H3.763V43.89h7.563c1.845 0 3.363.617 4.568 1.85 1.195 1.232 1.8 2.794 1.8 4.695zm-9.819-3.134v6.263h2.815c.83 0 1.509-.292 2.044-.88.535-.588.798-1.336.798-2.249s-.267-1.666-.798-2.25c-.53-.587-1.214-.879-2.044-.879H7.874v-.005zM31.315 62l-.61-2.197h-6.289L23.807 62H19.47l6.012-18.11h4.135L35.657 62h-4.342zm-5.985-5.486h4.467l-2.234-8.046-2.233 8.046zM39.79 62V47.617h-4.872V43.89H48.77v3.727h-4.896V62H39.79zm22.403-18.11h4.54L57.929 62h-4.568l3.3-6.832-6.295-11.278h4.771l3.78 7.143 3.276-7.143zM18.02 34.725c8.849 0 16.021-7.314 16.021-16.336 0-9.022-7.172-16.336-16.02-16.336C9.173 2.053 2 9.367 2 18.39c0 9.022 7.173 16.336 16.02 16.336z"/><path d="M24.808 14.287h-2.713v-1.473c0-1.91-1.523-3.463-3.396-3.463H16.69c-1.873 0-3.396 1.553-3.396 3.463v1.473h-2.713a.698.698 0 00-.692.705c0 .39.309.706.692.706h14.23a.09.09 0 01.065.028l-2.266 9.782a.75.75 0 00-.014.25.088.088 0 01-.023.075c-.018.023-.041.033-.074.033h-9.532a.698.698 0 00-.693.705.7.7 0 00.693.706H22.5c.425 0 .826-.183 1.108-.508a1.53 1.53 0 00.374-1.04l2.284-9.871a.752.752 0 00.014-.25 1.495 1.495 0 00-1.472-1.322zm-10.129-1.473c0-1.13.905-2.052 2.012-2.052h2.012c1.107 0 2.012.923 2.012 2.052v1.473h-6.036v-1.473z"/><path d="M14.006 23.744a.55.55 0 00-.545-.555h-3.105a.55.55 0 00-.544.555.55.55 0 00.544.555h3.105c.3 0 .545-.25.545-.555zm-3.77-5.18h3.105c.3 0 .545-.25.545-.556a.55.55 0 00-.545-.555h-3.105a.55.55 0 00-.544.555.55.55 0 00.544.556zm4.942 2.347a.551.551 0 00-.545-.555H7.731a.55.55 0 00-.545.555.55.55 0 00.545.555h6.898a.546.546 0 00.549-.555z"/></svg>
+<svg aria-hidden="true" width={size} height={size} id={id}  style={style} class={className} fill={fill} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 168 64"><path d="M78.24 16.645c0-6.117 2.883-9.21 8.547-9.21 5.664 0 8.525 3.087 8.525 9.21v18.12h5.814v-18.55c-.035-3.252-.617-5.552-1.91-7.71C97.677 6.082 94.009 2 86.788 2c-7.238 0-10.901 4.082-12.44 6.505-1.304 2.158-1.87 4.458-1.927 7.71v18.55h5.82v-18.12z"/><path fill-rule="evenodd" d="M53.046 2.017c-9.004 0-16.052 7.2-16.052 16.398 0 9.192 7.048 16.397 16.052 16.397 8.98 0 16.028-7.205 16.028-16.397C69.07 9.21 62.02 2.017 53.046 2.017zm0 27.178c-5.722 0-10.377-4.74-10.377-10.78 0-6.04 4.655-10.793 10.377-10.793 5.681 0 10.364 4.752 10.364 10.793 0 6.04-4.683 10.78-10.364 10.78zm82.553-11.215c-.052-8.781-7.019-15.927-15.537-15.927-8.768 0-15.643 7.17-15.643 16.32 0 4.3 1.615 8.375 4.522 11.493 2.959 3.152 6.864 4.887 11.017 4.887 4.741 0 11.379-2.317 14.541-8.893l.08-.211h-6.356l-.041.047c-1.707 2.076-4.925 3.482-7.993 3.482-3.9 0-8.572-2.618-9.852-8.523h25.141c.058-.611.121-1.835.121-2.675zm-25.078-2.541c.773-3.958 4.494-7.822 9.494-7.822 4.995 0 8.71 3.864 9.477 7.822h-18.971z"/><path d="M165.846 2.018l-16.962 42.169h-6.051l4.695-11.698-12.533-30.471h6.39l9.176 22.907 9.194-22.907h6.091z"/><path fill-rule="evenodd" d="M16.807 53.94c.59-1.026.886-2.192.886-3.505 0-1.9-.604-3.463-1.8-4.696-1.204-1.232-2.722-1.849-4.567-1.849H3.763V62h4.111v-5.02h2.155L12.821 62h4.67l-3.12-5.745a5.716 5.716 0 002.436-2.315zm-8.933-.376v-6.258h2.815c.83 0 1.513.292 2.044.88.53.583.798 1.336.798 2.249s-.263 1.66-.798 2.249c-.535.588-1.214.88-2.044.88H7.874zM31.315 62l-.61-2.197h-6.289L23.807 62H19.47l6.012-18.11h4.135L35.657 62h-4.342zm-5.985-5.486h4.467l-2.234-8.046-2.233 8.046z"/><path d="M39.79 47.617V62h4.084V47.617h4.896V43.89H34.918v3.727h4.873zm22.403-3.727h4.54L57.929 62h-4.568l3.3-6.832-6.295-11.278h4.771l3.78 7.143 3.276-7.143zM16.685 10.762c-1.107 0-2.011.922-2.011 2.051v1.473h6.035v-1.473c0-1.13-.904-2.051-2.012-2.051h-2.012z"/><path fill-rule="evenodd" d="M34.041 18.39c0 9.021-7.172 16.335-16.02 16.335C9.173 34.725 2 27.411 2 18.389 2 9.367 9.173 2.053 18.02 2.053c8.849 0 16.021 7.314 16.021 16.336zM22.09 14.285h2.713c.747 0 1.38.569 1.472 1.322a.75.75 0 01-.014.25l-2.284 9.87c.009.381-.125.753-.374 1.04a1.461 1.461 0 01-1.107.508h-9.533a.698.698 0 01-.692-.705c0-.39.309-.706.692-.706h9.533c.032 0 .055-.01.073-.033a.088.088 0 00.023-.075.752.752 0 01.014-.25l2.266-9.782a.09.09 0 00-.065-.028h-14.23a.698.698 0 01-.692-.705c0-.391.31-.706.692-.706h2.713v-1.473c0-1.91 1.523-3.463 3.396-3.463h2.008c1.873 0 3.395 1.553 3.395 3.463v1.473zm-8.633 8.902c.3 0 .544.25.544.555a.55.55 0 01-.544.555H10.35a.55.55 0 01-.544-.555.55.55 0 01.544-.555h3.106zm-3.226-4.625h3.105c.3 0 .545-.25.545-.555a.551.551 0 00-.545-.556H10.23a.55.55 0 00-.544.555.55.55 0 00.544.556zm4.397 1.792c.3 0 .545.25.545.555 0 .31-.245.56-.55.555H7.726a.55.55 0 01-.545-.555.55.55 0 01.545-.555h6.902z"/></svg>
