@@ -459,6 +459,13 @@ export const MIconPaymentMonochromeBizum24 = GenerateIcon('1.5rem', '0 0 72 24',
   ]
 ])
 
+export const MIconPaymentMonochromeBlik24 = GenerateIcon('1.5rem', '0 0 131 64', [
+  [
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M9.399 2h112.007a7.405 7.405 0 017.399 7.392v45.217A7.403 7.403 0 01121.406 62H9.399A7.403 7.403 0 012 54.609V9.392A7.403 7.403 0 019.399 2zm40.998 13.958a6.626 6.626 0 10-13.252-.086 6.626 6.626 0 0013.252.086zm-14.05 9.681a13.992 13.992 0 00-6.644 1.668V11.414h-7.408v28.272A14.051 14.051 0 1036.348 25.64zm0 20.81a6.763 6.763 0 11.001-13.525 6.763 6.763 0 010 13.525zm19.258-35.036h7.407v42.028h-7.407V11.413zm21.387 14.5h-7.407v27.528h7.407V25.914zm33.908 27.53h-9.541l-10.395-14.69.006 14.69h-7.41V11.42h7.41v27.298l10.208-12.801h8.654L99.437 38.638 110.9 53.443z"} },
+  ]
+])
+
 export const MIconPaymentMonochromeBonificoBancario24 = GenerateIcon('1.5rem', '0 0 78 24', [
   [
     'path',
@@ -533,13 +540,22 @@ export const MIconPaymentMonochromeOney4x24 = GenerateIcon('1.5rem', '0 0 65 24'
 export const MIconPaymentMonochromeOneyRaty24 = GenerateIcon('1.5rem', '0 0 60 24', [
   [
     'path',
-    { attrs: {"d":"M19.63 11.065c-1.906 0-3.458-1.58-3.458-3.594 0-2.013 1.552-3.597 3.459-3.597 1.893 0 3.455 1.584 3.455 3.597 0 2.014-1.561 3.594-3.455 3.594zm0-9.06c-3 0-5.35 2.4-5.35 5.466 0 3.065 2.35 5.466 5.35 5.466 2.994 0 5.344-2.401 5.344-5.466-.002-3.068-2.352-5.465-5.343-5.465zm27.519 5.321c-.018-2.926-2.34-5.308-5.18-5.308-2.922 0-5.214 2.39-5.214 5.44a5.59 5.59 0 001.508 3.83c.986 1.051 2.287 1.63 3.672 1.63 1.58 0 3.793-.773 4.847-2.965l.027-.07h-2.12l-.013.016c-.569.692-1.642 1.16-2.664 1.16-1.3 0-2.857-.872-3.284-2.84h8.38c.02-.204.04-.612.04-.893zm-8.36-.846c.258-1.32 1.498-2.608 3.165-2.608 1.665 0 2.903 1.288 3.159 2.608h-6.324zm-7.911-2.669c-1.888 0-2.85 1.032-2.85 3.07v6.04h-1.94V6.739c.02-1.084.208-1.85.643-2.57C27.244 3.361 28.465 2 30.878 2c2.407 0 3.63 1.36 4.143 2.168.43.72.625 1.486.636 2.57v6.184H33.72v-6.04c0-2.041-.953-3.07-2.841-3.07zM57.23 2.006l-5.653 14.056H49.56l1.565-3.9-4.178-10.156h2.13l3.059 7.636L55.2 2.006h2.03zM7.847 18.145c0 .438-.099.826-.296 1.168a1.9 1.9 0 01-.812.772L7.78 22H6.222l-.93-1.673h-.718V22h-1.37v-6.037h2.52c.615 0 1.121.206 1.523.617.398.41.6.931.6 1.565zm-3.273-1.044v2.087h.938a.879.879 0 00.681-.293c.179-.196.266-.446.266-.75s-.089-.555-.266-.75a.876.876 0 00-.681-.293h-.938V17.1zM12.387 22l-.203-.732h-2.096L9.885 22H8.439l2.004-6.037h1.378L13.834 22h-1.447zm-1.995-1.829h1.489l-.744-2.682-.745 2.682zM15.212 22v-4.794h-1.624v-1.243h4.617v1.243h-1.631V22h-1.362zm7.468-6.037h1.513L21.26 22h-1.523l1.1-2.277-2.098-3.76h1.59l1.26 2.381 1.092-2.38zM7.956 12.908c2.949 0 5.34-2.438 5.34-5.445s-2.391-5.445-5.34-5.445c-2.95 0-5.34 2.438-5.34 5.445s2.39 5.445 5.34 5.445z"} },
+    { attrs: {"d":"M28.028 6.881c0-2.038.962-3.07 2.85-3.07 1.888 0 2.841 1.03 2.841 3.07v6.04h1.938V6.739c-.011-1.084-.206-1.85-.636-2.57C34.507 3.361 33.285 2 30.878 2c-2.413 0-3.634 1.36-4.147 2.168-.435.72-.623 1.486-.643 2.57v6.184h1.94v-6.04z"} },
   ],[
     'path',
-    { attrs: {"d":"M10.218 6.095h-.904v-.49c0-.637-.508-1.155-1.132-1.155h-.67c-.624 0-1.131.518-1.131 1.155v.49h-.905a.233.233 0 00-.23.236c0 .13.102.235.23.235h4.744a.03.03 0 01.021.01l-.755 3.26a.25.25 0 00-.005.083.03.03 0 01-.007.025.03.03 0 01-.025.011H6.271a.233.233 0 00-.23.236c0 .13.103.235.23.235H9.45a.487.487 0 00.37-.17.51.51 0 00.124-.346l.761-3.29a.25.25 0 00.005-.084.498.498 0 00-.491-.44zm-3.376-.49c0-.377.301-.684.67-.684h.671c.37 0 .67.307.67.684v.49h-2.01v-.49z"} },
+    { attrs: {"fill-rule":"evenodd","d":"M19.63 2.006c-3 0-5.35 2.4-5.35 5.465s2.35 5.466 5.35 5.466c2.994 0 5.343-2.401 5.343-5.466-.002-3.068-2.351-5.465-5.342-5.465zm0 9.059c-1.907 0-3.458-1.58-3.458-3.594 0-2.013 1.551-3.597 3.459-3.597 1.893 0 3.454 1.584 3.454 3.597 0 2.014-1.56 3.594-3.454 3.594zm27.518-3.739c-.017-2.926-2.34-5.308-5.179-5.308-2.922 0-5.214 2.39-5.214 5.44 0 1.433.538 2.791 1.507 3.83.986 1.051 2.288 1.63 3.672 1.63 1.58 0 3.794-.773 4.847-2.965l.027-.07H44.69l-.014.016c-.57.692-1.642 1.16-2.665 1.16-1.3 0-2.857-.872-3.284-2.84h8.381c.02-.204.04-.612.04-.893zM38.79 6.48c.258-1.32 1.498-2.608 3.165-2.608 1.665 0 2.903 1.288 3.159 2.608H38.79z"} },
   ],[
     'path',
-    { attrs: {"d":"M6.617 9.248a.184.184 0 00-.181-.185H5.4c-.1 0-.182.083-.182.185s.082.185.182.185h1.035c.1 0 .181-.083.181-.185zM5.36 7.521h1.036c.1 0 .181-.083.181-.185a.184.184 0 00-.181-.185H5.36c-.1 0-.182.083-.182.185s.082.185.182.185zm1.648.783a.184.184 0 00-.181-.185H4.526c-.1 0-.182.083-.182.185s.082.185.182.185h2.3a.182.182 0 00.182-.185z"} },
+    { attrs: {"d":"M57.23 2.006l-5.654 14.056H49.56l1.566-3.9-4.178-10.156h2.13l3.059 7.636L55.2 2.006h2.03z"} },
+  ],[
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M7.551 19.313c.197-.341.295-.73.295-1.168 0-.634-.201-1.154-.6-1.565-.401-.411-.907-.617-1.522-.617H3.203V22h1.37v-1.673h.719L6.222 22H7.78l-1.04-1.915c.344-.172.615-.43.812-.772zm-2.978-.125v-2.086h.939c.276 0 .504.097.68.293.178.195.267.446.267.75s-.088.554-.266.75a.879.879 0 01-.681.293h-.939zM12.387 22l-.203-.732h-2.097L9.884 22H8.44l2.004-6.037h1.378L13.834 22h-1.447zm-1.995-1.829h1.489l-.745-2.682-.744 2.682z"} },
+  ],[
+    'path',
+    { attrs: {"d":"M15.212 17.206V22h1.361v-4.794h1.632v-1.243h-4.617v1.243h1.624zm7.468-1.243h1.513L21.258 22h-1.522l1.1-2.277-2.099-3.76h1.59l1.26 2.381 1.092-2.38zM7.51 4.92a.679.679 0 00-.67.684v.491h2.012v-.49a.679.679 0 00-.671-.684h-.67z"} },
+  ],[
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M13.296 7.463c0 3.007-2.391 5.445-5.34 5.445-2.95 0-5.34-2.438-5.34-5.445s2.39-5.445 5.34-5.445c2.949 0 5.34 2.438 5.34 5.445zM9.31 6.095h.905c.249 0 .46.19.49.44a.25.25 0 01-.004.084L9.94 9.91a.51.51 0 01-.124.347.487.487 0 01-.37.17H6.27a.233.233 0 01-.23-.236c0-.13.102-.235.23-.235h3.178c.01 0 .018-.003.024-.011a.03.03 0 00.008-.025.25.25 0 01.005-.083l.755-3.26a.03.03 0 00-.022-.01H5.474a.233.233 0 01-.23-.236c0-.13.102-.235.23-.235h.904v-.49c0-.637.508-1.155 1.132-1.155h.67c.624 0 1.131.518 1.131 1.154v.491zM6.434 9.063c.1 0 .181.083.181.185a.184.184 0 01-.181.185H5.399a.184.184 0 01-.182-.185c0-.102.082-.185.182-.185h1.035zM5.359 7.52h1.035c.1 0 .181-.083.181-.185a.184.184 0 00-.181-.185H5.359c-.1 0-.182.083-.182.185s.082.185.182.185zm1.465.597c.1 0 .182.083.182.185a.182.182 0 01-.183.185h-2.3a.184.184 0 01-.181-.185c0-.102.081-.185.181-.185h2.301z"} },
   ]
 ])
 
@@ -1658,6 +1674,13 @@ export const MIconSafePayment224 = GenerateIcon('1.5rem', '0 0 24 24', [
   ],[
     'path',
     { attrs: {"d":"M14.15 13.52l-2.84 2.79L9.83 15a.5.5 0 10-.66.74L11 17.37a.5.5 0 00.34.13.54.54 0 00.35-.14l3.17-3.13a.5.5 0 10-.7-.71z"} },
+  ]
+])
+
+export const MIconTransfer24 = GenerateIcon('1.5rem', '0 0 24 24', [
+  [
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M17.208 2.292a1 1 0 10-1.415 1.414L18.586 6.5H5a1 1 0 000 2h13.586l-2.793 2.793a1 1 0 101.415 1.414l4.5-4.5a1 1 0 000-1.414l-4.5-4.5zm-9 10.414a1 1 0 10-1.414-1.414l-4.5 4.5a1 1 0 000 1.414l4.5 4.5a1 1 0 101.414-1.414l-2.793-2.793h13.586a1 1 0 100-2H5.415l2.793-2.793z"} },
   ]
 ])
 
@@ -3289,6 +3312,16 @@ export const MIconPaymentMonochromeBizum32 = GenerateIcon('2rem', '0 0 99 32', [
   ]
 ])
 
+export const MIconPaymentMonochromeBlik32 = GenerateIcon('2rem', '0 0 98 48', [
+  [
+    'path',
+    { attrs: {"d":"M28.096 34.596a4.959 4.959 0 110-9.918 4.959 4.959 0 010 9.918z"} },
+  ],[
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M8.333 2H90.47a5.43 5.43 0 015.427 5.42v33.16A5.43 5.43 0 0190.47 46H8.333a5.43 5.43 0 01-5.426-5.42V7.42A5.43 5.43 0 018.333 2zm30.066 10.236a4.86 4.86 0 10-9.719-.063 4.86 4.86 0 009.719.063zm-10.303 7.1c-1.7-.003-3.375.417-4.873 1.222V8.904h-5.432v20.733a10.304 10.304 0 1010.305-10.302zM42.218 8.902h5.432v30.82h-5.432V8.904zm15.684 10.634H52.47v20.187h5.432V19.537zm24.866 20.188H75.77l-7.623-10.773.004 10.773h-5.433V8.907h5.433v20.019l7.486-9.388h6.346l-7.623 9.33 8.407 10.857z"} },
+  ]
+])
+
 export const MIconPaymentMonochromeBonificoBancario32 = GenerateIcon('2rem', '0 0 108 33', [
   [
     'path',
@@ -3363,13 +3396,22 @@ export const MIconPaymentMonochromeOney4x32 = GenerateIcon('2rem', '0 0 89 32', 
 export const MIconPaymentMonochromeOneyRaty32 = GenerateIcon('2rem', '0 0 82 32', [
   [
     'path',
-    { attrs: {"d":"M26.514 14.691c-2.67 0-4.842-2.212-4.842-5.03 0-2.82 2.172-5.037 4.842-5.037 2.651 0 4.837 2.217 4.837 5.036 0 2.819-2.186 5.031-4.837 5.031zm0-12.683c-4.202 0-7.491 3.36-7.491 7.652 0 4.29 3.29 7.652 7.49 7.652 4.192 0 7.48-3.362 7.48-7.652-.002-4.295-3.291-7.652-7.48-7.652zm38.525 7.449c-.025-4.098-3.276-7.432-7.251-7.432-4.092 0-7.3 3.345-7.3 7.616 0 2.006.754 3.908 2.11 5.363 1.381 1.471 3.203 2.28 5.141 2.28 2.213 0 5.31-1.08 6.786-4.15l.038-.098h-2.967l-.018.022c-.797.969-2.3 1.625-3.731 1.625-1.82 0-4-1.222-4.597-3.977h11.732a16.3 16.3 0 00.057-1.249zM53.336 8.271c.36-1.847 2.096-3.65 4.43-3.65 2.331 0 4.065 1.803 4.423 3.65h-8.853zM42.26 4.536c-2.643 0-3.99 1.444-3.99 4.298v8.456h-2.715V8.634c.027-1.518.29-2.591.899-3.598C37.172 3.905 38.882 2 42.26 2c3.37 0 5.082 1.905 5.8 3.036.603 1.007.875 2.08.891 3.598v8.656h-2.713V8.834c0-2.857-1.335-4.298-3.978-4.298zm36.894-2.528l-7.916 19.68h-2.824l2.191-5.46-5.848-14.22h2.982l4.282 10.69 4.29-10.69h2.843zM10.016 24.603c0 .613-.138 1.157-.414 1.636a2.668 2.668 0 01-1.136 1.08L9.92 30H7.742l-1.303-2.343H5.434V30H3.515v-8.451h3.53c.86 0 1.57.287 2.131.863.558.575.84 1.304.84 2.191zm-4.582-1.462v2.922h1.313c.388 0 .704-.136.954-.41.25-.275.373-.624.373-1.05 0-.426-.125-.777-.373-1.05a1.227 1.227 0 00-.954-.41H5.434v-.002zM16.372 30l-.284-1.025h-2.935L12.87 30h-2.024l2.806-8.451h1.93L18.398 30h-2.027zm-2.792-2.56h2.084l-1.042-3.755-1.042 3.755zM20.328 30v-6.712h-2.274v-1.74h6.464v1.74h-2.284V30h-1.906zm10.454-8.451h2.12L28.791 30H26.66l1.54-3.188-2.938-5.263h2.227l1.763 3.333 1.53-3.333zm-20.613-4.277c4.129 0 7.476-3.413 7.476-7.624 0-4.21-3.347-7.623-7.476-7.623-4.13 0-7.477 3.413-7.477 7.623 0 4.21 3.348 7.624 7.477 7.624z"} },
+    { attrs: {"d":"M38.27 8.834c0-2.854 1.347-4.298 3.99-4.298 2.643 0 3.978 1.441 3.978 4.298v8.456h2.713V8.634c-.016-1.518-.288-2.591-.89-3.598C47.34 3.905 45.63 2 42.26 2c-3.378 0-5.087 1.905-5.806 3.036-.608 1.007-.872 2.08-.9 3.598v8.656h2.717V8.834z"} },
   ],[
     'path',
-    { attrs: {"d":"M13.336 7.734H12.07v-.688c0-.891-.71-1.616-1.585-1.616h-.937c-.874 0-1.585.725-1.585 1.616v.688H6.697a.326.326 0 00-.323.33c0 .181.145.328.323.328h6.641c.01 0 .022.005.03.014l-1.057 4.564a.351.351 0 00-.006.117.04.04 0 01-.011.035.044.044 0 01-.035.015H7.811a.326.326 0 00-.323.33c0 .182.144.329.323.329h4.448a.682.682 0 00.517-.237.715.715 0 00.175-.485l1.065-4.607a.35.35 0 00.007-.116.698.698 0 00-.687-.617zm-4.727-.688a.95.95 0 01.94-.957h.938a.95.95 0 01.939.957v.688H8.609v-.688z"} },
+    { attrs: {"fill-rule":"evenodd","d":"M26.514 2.008c-4.202 0-7.491 3.36-7.491 7.652 0 4.29 3.29 7.652 7.49 7.652 4.192 0 7.48-3.362 7.48-7.652-.002-4.295-3.291-7.652-7.48-7.652zm0 12.683c-2.67 0-4.842-2.212-4.842-5.03 0-2.82 2.172-5.037 4.842-5.037 2.651 0 4.837 2.217 4.837 5.036 0 2.819-2.186 5.031-4.837 5.031zm38.525-5.234c-.025-4.098-3.276-7.432-7.251-7.432-4.092 0-7.3 3.345-7.3 7.616 0 2.006.754 3.908 2.11 5.363 1.381 1.471 3.203 2.28 5.141 2.28 2.213 0 5.31-1.08 6.786-4.15l.038-.098h-2.967l-.018.022c-.797.969-2.3 1.625-3.731 1.625-1.82 0-4-1.222-4.597-3.977h11.732a16.3 16.3 0 00.057-1.249zM53.336 8.271c.36-1.847 2.096-3.65 4.43-3.65 2.331 0 4.065 1.803 4.423 3.65h-8.853z"} },
   ],[
     'path',
-    { attrs: {"d":"M8.295 12.147a.257.257 0 00-.254-.259h-1.45a.257.257 0 00-.253.26c0 .142.114.258.254.258H8.04c.14 0 .254-.116.254-.259zM6.536 9.73h1.449c.14 0 .254-.117.254-.26a.257.257 0 00-.254-.259h-1.45a.257.257 0 00-.253.26c0 .142.114.259.254.259zm2.306 1.095a.257.257 0 00-.254-.259H5.367a.257.257 0 00-.254.26c0 .142.114.258.254.258h3.219a.255.255 0 00.256-.259z"} },
+    { attrs: {"d":"M79.154 2.008l-7.916 19.68h-2.824l2.191-5.46-5.848-14.22h2.982l4.282 10.69 4.29-10.69h2.843z"} },
+  ],[
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M9.602 26.239c.276-.479.414-1.023.414-1.636 0-.887-.282-1.616-.84-2.191-.562-.576-1.27-.863-2.132-.863H3.515V30h1.919v-2.343h1.005L7.742 30h2.18l-1.457-2.68a2.668 2.668 0 001.137-1.081zm-4.168-.176v-2.92h1.313c.388 0 .706.136.954.41.248.273.373.624.373 1.05 0 .426-.123.775-.373 1.05a1.23 1.23 0 01-.954.41H5.434zM16.372 30l-.284-1.025h-2.935L12.87 30h-2.024l2.806-8.451h1.93L18.398 30h-2.027zm-2.792-2.56h2.084l-1.042-3.755-1.042 3.755z"} },
+  ],[
+    'path',
+    { attrs: {"d":"M20.328 23.288V30h1.906v-6.712h2.284v-1.74h-6.464v1.74h2.274zm10.454-1.739h2.12L28.791 30H26.66l1.54-3.188-2.938-5.263h2.227l1.763 3.333 1.53-3.333zM9.546 6.089a.95.95 0 00-.94.957v.687h2.817v-.687a.95.95 0 00-.939-.957h-.938z"} },
+  ],[
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M17.645 9.648c0 4.21-3.347 7.624-7.476 7.624-4.13 0-7.477-3.413-7.477-7.624 0-4.21 3.348-7.623 7.477-7.623 4.129 0 7.476 3.413 7.476 7.623zm-5.578-1.915h1.266c.349 0 .644.266.687.617a.35.35 0 01-.006.117l-1.066 4.606a.714.714 0 01-.175.486.682.682 0 01-.516.237h-4.45a.326.326 0 01-.322-.33c0-.182.144-.33.323-.33h4.449c.015 0 .025-.003.034-.014a.04.04 0 00.01-.036.35.35 0 01.007-.116l1.058-4.565a.042.042 0 00-.03-.013H6.694a.326.326 0 01-.323-.33c0-.182.144-.329.323-.329H7.96v-.687c0-.891.71-1.616 1.585-1.616h.936c.874 0 1.585.725 1.585 1.616v.687zm-4.029 4.155c.14 0 .254.116.254.259a.257.257 0 01-.254.259H6.59a.257.257 0 01-.254-.26c0-.142.114-.258.254-.258h1.45zM6.533 9.729h1.45c.14 0 .253-.116.253-.259a.257.257 0 00-.254-.259H6.533a.257.257 0 00-.254.26c0 .142.114.258.254.258zm2.052.837c.14 0 .254.116.254.259a.255.255 0 01-.256.259h-3.22a.257.257 0 01-.253-.26c0-.142.114-.258.254-.258h3.221z"} },
   ]
 ])
 
@@ -4456,6 +4498,13 @@ export const MIconSafePayment232 = GenerateIcon('2rem', '0 0 32 32', [
   ],[
     'path',
     { attrs: {"d":"M14.73 22.44a.51.51 0 00.34.14.51.51 0 00.36-.15L19.8 18a.5.5 0 000-.71.49.49 0 00-.7 0l-4 4.09-2.16-2a.5.5 0 00-.71 0 .5.5 0 000 .71z"} },
+  ]
+])
+
+export const MIconTransfer32 = GenerateIcon('2rem', '0 0 32 32', [
+  [
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M22.707 3.292a1 1 0 10-1.414 1.414L25.586 9H6.667a1 1 0 000 2h18.92l-4.294 4.293a1 1 0 001.414 1.414l6-6a1 1 0 000-1.414l-6-6zM10.708 16.706a1 1 0 00-1.414-1.414l-6 6a1 1 0 000 1.414l6 6a1 1 0 101.414-1.414l-4.293-4.293h18.92a1 1 0 100-2H6.414l4.293-4.293z"} },
   ]
 ])
 
@@ -6114,6 +6163,16 @@ export const MIconPaymentMonochromeBizum48 = GenerateIcon('3rem', '0 0 153 48', 
   ]
 ])
 
+export const MIconPaymentMonochromeBlik48 = GenerateIcon('3rem', '0 0 64 32', [
+  [
+    'path',
+    { attrs: {"d":"M18.843 22.743a3.155 3.155 0 110-6.31 3.155 3.155 0 010 6.31z"} },
+  ],[
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M6.267 2h52.27a3.455 3.455 0 013.453 3.45v21.1A3.456 3.456 0 0158.537 30H6.267a3.455 3.455 0 01-3.453-3.45V5.45A3.455 3.455 0 016.267 2zM25.4 8.514a3.092 3.092 0 10-6.184-.04 3.092 3.092 0 006.184.04zm-6.556 4.518a6.53 6.53 0 00-3.1.778V6.393h-3.458v13.194a6.557 6.557 0 106.558-6.555zm8.987-6.64h3.457v19.614H27.83V6.393zm9.98 6.768h-3.456v12.846h3.457V13.16zm15.824 12.847h-4.452L44.33 19.15l.002 6.856h-3.457V6.395h3.457v12.74l4.764-5.974h4.039l-4.851 5.937 5.35 6.909z"} },
+  ]
+])
+
 export const MIconPaymentMonochromeBonificoBancario48 = GenerateIcon('3rem', '0 0 166 49', [
   [
     'path',
@@ -6188,13 +6247,22 @@ export const MIconPaymentMonochromeOney4x48 = GenerateIcon('3rem', '0 0 138 48',
 export const MIconPaymentMonochromeOneyRaty48 = GenerateIcon('3rem', '0 0 125 48', [
   [
     'path',
-    { attrs: {"d":"M40.28 21.943c-4.196 0-7.61-3.476-7.61-7.906s3.414-7.914 7.61-7.914c4.166 0 7.6 3.485 7.6 7.914 0 4.43-3.434 7.906-7.6 7.906zm0-19.93c-6.603 0-11.772 5.279-11.772 12.024 0 6.741 5.17 12.025 11.772 12.025 6.585 0 11.754-5.284 11.754-12.025-.004-6.75-5.173-12.024-11.754-12.024zm60.539 11.705c-.038-6.439-5.147-11.68-11.394-11.68-6.43 0-11.471 5.258-11.471 11.97 0 3.152 1.184 6.14 3.316 8.427 2.17 2.312 5.033 3.584 8.079 3.584 3.476 0 8.344-1.7 10.663-6.521l.059-.156H95.41l-.03.035c-1.252 1.522-3.612 2.553-5.862 2.553-2.86 0-6.286-1.92-7.225-6.25h18.437a25.82 25.82 0 00.089-1.962zm-18.39-1.863c.566-2.903 3.294-5.736 6.962-5.736 3.663 0 6.387 2.833 6.95 5.736H82.428zm-17.406-5.87c-4.153 0-6.268 2.269-6.268 6.754v13.288h-4.268V12.425c.043-2.385.457-4.071 1.413-5.654C57.03 4.993 59.715 2 65.023 2c5.296 0 7.986 2.993 9.115 4.77.948 1.583 1.375 3.27 1.4 5.654v13.604h-4.263V12.739c0-4.49-2.098-6.754-6.252-6.754zM123 2.013l-12.439 30.924h-4.437l3.443-8.579-9.191-22.345h4.686l6.729 16.799 6.742-16.8H123zM14.355 37.519c0 .963-.217 1.818-.65 2.57a4.192 4.192 0 01-1.787 1.698L14.206 46H10.78l-2.047-3.681h-1.58V46H4.139V32.72h5.546c1.354 0 2.467.452 3.35 1.355.876.904 1.32 2.05 1.32 3.444zm-7.201-2.298v4.593h2.064c.61 0 1.107-.215 1.499-.646.393-.431.585-.98.585-1.649 0-.67-.196-1.221-.585-1.65-.39-.43-.89-.645-1.499-.645H7.154v-.003zM24.344 46l-.447-1.611h-4.612L18.838 46h-3.18l4.409-13.28h3.031L27.528 46h-3.184zm-4.39-4.023h3.276l-1.637-5.9-1.638 5.9zM30.56 46V35.452h-3.574V32.72h10.158v2.733h-3.59V46h-2.995zm16.427-13.28h3.33L43.861 46h-3.35l2.42-5.01-4.616-8.27h3.499l2.771 5.237 2.402-5.238zm-32.392-6.722c6.488 0 11.748-5.363 11.748-11.98 0-6.616-5.26-11.979-11.748-11.979-6.489 0-11.749 5.363-11.749 11.98 0 6.616 5.26 11.98 11.749 11.98z"} },
+    { attrs: {"d":"M58.755 12.74c0-4.486 2.114-6.755 6.268-6.755s6.251 2.264 6.251 6.754v13.288h4.264V12.425c-.025-2.385-.453-4.071-1.4-5.654C73.008 4.993 70.318 2 65.023 2c-5.308 0-7.994 2.993-9.123 4.77-.956 1.583-1.37 3.27-1.413 5.654v13.604h4.267V12.739z"} },
   ],[
     'path',
-    { attrs: {"d":"M19.572 11.01h-1.99V9.93c0-1.4-1.117-2.54-2.49-2.54H13.62c-1.374 0-2.49 1.14-2.49 2.54v1.08H9.14a.512.512 0 00-.508.518c0 .286.227.517.508.517h10.435a.07.07 0 01.048.02L17.96 19.24a.55.55 0 00-.01.183.064.064 0 01-.017.055.063.063 0 01-.054.024h-6.99a.512.512 0 00-.508.518c0 .286.226.518.507.518h6.99c.312 0 .607-.135.813-.373.183-.21.28-.483.274-.763l1.675-7.239a.55.55 0 00.01-.182 1.096 1.096 0 00-1.08-.97zm-7.428-1.08c0-.828.664-1.504 1.476-1.504h1.475c.812 0 1.475.676 1.475 1.504v1.08h-4.426V9.93z"} },
+    { attrs: {"fill-rule":"evenodd","d":"M40.28 2.013c-6.603 0-11.772 5.279-11.772 12.024 0 6.741 5.169 12.025 11.771 12.025 6.586 0 11.755-5.284 11.755-12.025-.005-6.75-5.173-12.024-11.755-12.024zm0 19.93c-4.197 0-7.61-3.476-7.61-7.906s3.413-7.914 7.61-7.914c4.166 0 7.6 3.485 7.6 7.914 0 4.43-3.434 7.906-7.6 7.906zm60.539-8.225c-.039-6.439-5.148-11.68-11.395-11.68-6.43 0-11.47 5.258-11.47 11.97 0 3.152 1.183 6.14 3.315 8.427 2.17 2.312 5.033 3.584 8.08 3.584 3.476 0 8.344-1.7 10.662-6.521l.06-.156h-4.662l-.03.035c-1.252 1.522-3.612 2.553-5.862 2.553-2.86 0-6.285-1.92-7.224-6.25h18.437a25.82 25.82 0 00.089-1.962zm-18.39-1.863c.566-2.903 3.294-5.736 6.961-5.736 3.663 0 6.387 2.833 6.95 5.736H82.428z"} },
   ],[
     'path',
-    { attrs: {"d":"M11.65 17.945a.404.404 0 00-.399-.407H8.974c-.22 0-.4.183-.4.407 0 .225.18.408.4.408h2.277c.22 0 .4-.183.4-.408zm-2.764-3.798h2.277c.22 0 .4-.183.4-.408a.404.404 0 00-.4-.407H8.886c-.22 0-.4.183-.4.407 0 .225.18.408.4.408zm3.624 1.721a.404.404 0 00-.4-.407H7.05c-.22 0-.4.183-.4.407 0 .224.18.407.4.407h5.058a.4.4 0 00.403-.407z"} },
+    { attrs: {"d":"M123 2.013l-12.44 30.924h-4.437l3.443-8.579-9.191-22.345h4.686l6.73 16.799 6.742-16.8H123z"} },
+  ],[
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M13.704 40.09c.433-.753.65-1.608.65-2.571 0-1.394-.443-2.54-1.32-3.444-.883-.904-1.996-1.356-3.35-1.356H4.138V46h3.015v-3.681h1.58L10.782 46h3.424l-2.287-4.213a4.191 4.191 0 001.786-1.697zm-6.55-.277v-4.589h2.064c.609 0 1.11.214 1.498.646.39.428.586.98.586 1.649 0 .67-.193 1.218-.585 1.65-.393.43-.89.645-1.5.645H7.154zM24.343 46l-.447-1.611h-4.612L18.838 46h-3.181l4.409-13.28h3.032L27.528 46h-3.185zm-4.389-4.023h3.276l-1.638-5.9-1.638 5.9z"} },
+  ],[
+    'path',
+    { attrs: {"d":"M30.559 35.452V46h2.994V35.452h3.59V32.72H26.986v2.733h3.574zm16.428-2.732h3.33L43.86 46h-3.35l2.42-5.01-4.616-8.27h3.5l2.77 5.237 2.403-5.238zM13.615 8.425c-.812 0-1.475.676-1.475 1.505v1.08h4.426V9.93c0-.829-.664-1.505-1.476-1.505h-1.475z"} },
+  ],[
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M26.343 14.019c0 6.616-5.26 11.98-11.749 11.98-6.488 0-11.748-5.364-11.748-11.98 0-6.617 5.26-11.98 11.748-11.98 6.489 0 11.749 5.363 11.749 11.98zm-8.766-3.01h1.99c.548 0 1.012.418 1.08.97a.548.548 0 01-.01.183l-1.676 7.239c.007.28-.091.552-.274.762a1.07 1.07 0 01-.812.373h-6.99a.512.512 0 01-.508-.518c0-.286.226-.517.507-.517h6.991c.024 0 .04-.007.054-.024a.064.064 0 00.017-.055.552.552 0 01.01-.183l1.662-7.174a.066.066 0 00-.048-.02H9.135a.512.512 0 01-.508-.518c0-.286.227-.518.508-.518h1.99V9.93c0-1.4 1.116-2.539 2.49-2.539h1.472c1.374 0 2.49 1.139 2.49 2.54v1.08zm-6.33 6.529c.22 0 .399.183.399.407 0 .224-.18.407-.4.407H8.97c-.22 0-.4-.183-.4-.407 0-.224.18-.407.4-.407h2.277zM8.88 14.146h2.277c.22 0 .4-.183.4-.407a.404.404 0 00-.4-.407H8.881c-.22 0-.4.183-.4.407 0 .224.18.407.4.407zm3.225 1.314c.22 0 .4.183.4.407a.4.4 0 01-.404.408H7.044c-.22 0-.4-.183-.4-.407 0-.225.18-.408.4-.408h5.062z"} },
   ]
 ])
 
@@ -7272,6 +7340,13 @@ export const MIconSafePayment248 = GenerateIcon('3rem', '0 0 48 48', [
   ],[
     'path',
     { attrs: {"d":"M28.64 26.08l-6.11 6.21-3.2-3a1 1 0 00-1.33 1.5l3.92 3.63a1 1 0 00.68.27 1 1 0 00.71-.3l6.8-6.91a1 1 0 00-1.43-1.4z"} },
+  ]
+])
+
+export const MIconTransfer48 = GenerateIcon('3rem', '0 0 48 48', [
+  [
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M34.062 4.937A1.5 1.5 0 1031.94 7.06l6.44 6.439H10a1.5 1.5 0 000 3h28.38l-6.44 6.44a1.5 1.5 0 102.122 2.12l9-9a1.5 1.5 0 000-2.12l-9-9zm-18 20.122a1.5 1.5 0 10-2.12-2.122l-9 9a1.5 1.5 0 000 2.122l9 9a1.5 1.5 0 102.12-2.122l-6.439-6.439h28.379a1.5 1.5 0 100-3H9.623l6.44-6.44z"} },
   ]
 ])
 
@@ -8909,6 +8984,13 @@ export const MIconPaymentMonochromeBizum64 = GenerateIcon('4rem', '0 0 207 64', 
   ]
 ])
 
+export const MIconPaymentMonochromeBlik64 = GenerateIcon('4rem', '0 0 47 24', [
+  [
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M4.735 2H42.07a2.468 2.468 0 012.467 2.464v15.072A2.468 2.468 0 0142.07 22H4.735a2.468 2.468 0 01-2.466-2.464V4.464A2.468 2.468 0 014.735 2zM18.4 6.653a2.209 2.209 0 10-4.417-.03 2.209 2.209 0 004.417.03zM13.718 9.88a4.664 4.664 0 00-2.215.556V5.138h-2.47v9.424a4.683 4.683 0 104.685-4.682zm0 6.936a2.254 2.254 0 110-4.508 2.254 2.254 0 010 4.508zm6.42-11.678h2.468v14.01h-2.469V5.137zm7.128 4.833h-2.469v9.176h2.47V9.971zm11.303 9.177h-3.18l-3.466-4.897.002 4.897h-2.47V5.14h2.47v9.099l3.403-4.267h2.884l-3.464 4.24 3.82 4.936z"} },
+  ]
+])
+
 export const MIconPaymentMonochromeBonificoBancario64 = GenerateIcon('4rem', '0 0 225 65', [
   [
     'path',
@@ -8983,13 +9065,22 @@ export const MIconPaymentMonochromeOney4x64 = GenerateIcon('4rem', '0 0 186 64',
 export const MIconPaymentMonochromeOneyRaty64 = GenerateIcon('4rem', '0 0 168 64', [
   [
     'path',
-    { attrs: {"d":"M53.046 29.195c-5.722 0-10.377-4.74-10.377-10.78 0-6.04 4.655-10.793 10.377-10.793 5.681 0 10.364 4.752 10.364 10.793 0 6.04-4.683 10.78-10.364 10.78zm0-27.178c-9.004 0-16.052 7.2-16.052 16.398 0 9.192 7.048 16.397 16.052 16.397 8.98 0 16.028-7.205 16.028-16.397C69.07 9.21 62.02 2.017 53.046 2.017zm82.553 15.963c-.052-8.781-7.019-15.927-15.537-15.927-8.768 0-15.643 7.17-15.643 16.32 0 4.3 1.615 8.375 4.522 11.493 2.959 3.152 6.864 4.887 11.017 4.887 4.741 0 11.379-2.317 14.541-8.893l.08-.211h-6.356l-.041.047c-1.707 2.076-4.925 3.481-7.993 3.481-3.9 0-8.572-2.617-9.852-8.521h25.141a34.9 34.9 0 00.121-2.676zm-25.078-2.541c.773-3.958 4.494-7.822 9.494-7.822 4.995 0 8.71 3.864 9.477 7.822h-18.971zM86.787 7.434c-5.664 0-8.548 3.094-8.548 9.21v18.12h-5.82V16.216c.058-3.252.624-5.552 1.927-7.71C75.886 6.082 79.55 2 86.787 2c7.221 0 10.89 4.082 12.43 6.505 1.292 2.158 1.874 4.458 1.909 7.71v18.55h-5.814v-18.12c0-6.123-2.86-9.21-8.525-9.21zm79.059-5.416l-16.962 42.169h-6.051l4.695-11.698-12.533-30.471h6.39l9.176 22.907 9.194-22.907h6.091zM17.693 50.435c0 1.313-.295 2.48-.886 3.505a5.716 5.716 0 01-2.436 2.315L17.491 62h-4.67l-2.792-5.02H7.874V62H3.763V43.89h7.563c1.845 0 3.363.617 4.568 1.85 1.195 1.232 1.8 2.794 1.8 4.695zm-9.819-3.134v6.263h2.815c.83 0 1.509-.292 2.044-.88.535-.588.798-1.336.798-2.249s-.267-1.666-.798-2.25c-.53-.587-1.214-.879-2.044-.879H7.874v-.005zM31.315 62l-.61-2.197h-6.289L23.807 62H19.47l6.012-18.11h4.135L35.657 62h-4.342zm-5.985-5.486h4.467l-2.234-8.046-2.233 8.046zM39.79 62V47.617h-4.872V43.89H48.77v3.727h-4.896V62H39.79zm22.403-18.11h4.54L57.929 62h-4.568l3.3-6.832-6.295-11.278h4.771l3.78 7.143 3.276-7.143zM18.02 34.725c8.849 0 16.021-7.314 16.021-16.336 0-9.022-7.172-16.336-16.02-16.336C9.173 2.053 2 9.367 2 18.39c0 9.022 7.173 16.336 16.02 16.336z"} },
+    { attrs: {"d":"M78.24 16.645c0-6.117 2.883-9.21 8.547-9.21 5.664 0 8.525 3.087 8.525 9.21v18.12h5.814v-18.55c-.035-3.252-.617-5.552-1.91-7.71C97.677 6.082 94.009 2 86.788 2c-7.238 0-10.901 4.082-12.44 6.505-1.304 2.158-1.87 4.458-1.927 7.71v18.55h5.82v-18.12z"} },
   ],[
     'path',
-    { attrs: {"d":"M24.808 14.287h-2.713v-1.473c0-1.91-1.523-3.463-3.396-3.463H16.69c-1.873 0-3.396 1.553-3.396 3.463v1.473h-2.713a.698.698 0 00-.692.705c0 .39.309.706.692.706h14.23a.09.09 0 01.065.028l-2.266 9.782a.75.75 0 00-.014.25.088.088 0 01-.023.075c-.018.023-.041.033-.074.033h-9.532a.698.698 0 00-.693.705.7.7 0 00.693.706H22.5c.425 0 .826-.183 1.108-.508a1.53 1.53 0 00.374-1.04l2.284-9.871a.752.752 0 00.014-.25 1.495 1.495 0 00-1.472-1.322zm-10.129-1.473c0-1.13.905-2.052 2.012-2.052h2.012c1.107 0 2.012.923 2.012 2.052v1.473h-6.036v-1.473z"} },
+    { attrs: {"fill-rule":"evenodd","d":"M53.046 2.017c-9.004 0-16.052 7.2-16.052 16.398 0 9.192 7.048 16.397 16.052 16.397 8.98 0 16.028-7.205 16.028-16.397C69.07 9.21 62.02 2.017 53.046 2.017zm0 27.178c-5.722 0-10.377-4.74-10.377-10.78 0-6.04 4.655-10.793 10.377-10.793 5.681 0 10.364 4.752 10.364 10.793 0 6.04-4.683 10.78-10.364 10.78zm82.553-11.215c-.052-8.781-7.019-15.927-15.537-15.927-8.768 0-15.643 7.17-15.643 16.32 0 4.3 1.615 8.375 4.522 11.493 2.959 3.152 6.864 4.887 11.017 4.887 4.741 0 11.379-2.317 14.541-8.893l.08-.211h-6.356l-.041.047c-1.707 2.076-4.925 3.482-7.993 3.482-3.9 0-8.572-2.618-9.852-8.523h25.141c.058-.611.121-1.835.121-2.675zm-25.078-2.541c.773-3.958 4.494-7.822 9.494-7.822 4.995 0 8.71 3.864 9.477 7.822h-18.971z"} },
   ],[
     'path',
-    { attrs: {"d":"M14.006 23.744a.55.55 0 00-.545-.555h-3.105a.55.55 0 00-.544.555.55.55 0 00.544.555h3.105c.3 0 .545-.25.545-.555zm-3.77-5.18h3.105c.3 0 .545-.25.545-.556a.55.55 0 00-.545-.555h-3.105a.55.55 0 00-.544.555.55.55 0 00.544.556zm4.942 2.347a.551.551 0 00-.545-.555H7.731a.55.55 0 00-.545.555.55.55 0 00.545.555h6.898a.546.546 0 00.549-.555z"} },
+    { attrs: {"d":"M165.846 2.018l-16.962 42.169h-6.051l4.695-11.698-12.533-30.471h6.39l9.176 22.907 9.194-22.907h6.091z"} },
+  ],[
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M16.807 53.94c.59-1.026.886-2.192.886-3.505 0-1.9-.604-3.463-1.8-4.696-1.204-1.232-2.722-1.849-4.567-1.849H3.763V62h4.111v-5.02h2.155L12.821 62h4.67l-3.12-5.745a5.716 5.716 0 002.436-2.315zm-8.933-.376v-6.258h2.815c.83 0 1.513.292 2.044.88.53.583.798 1.336.798 2.249s-.263 1.66-.798 2.249c-.535.588-1.214.88-2.044.88H7.874zM31.315 62l-.61-2.197h-6.289L23.807 62H19.47l6.012-18.11h4.135L35.657 62h-4.342zm-5.985-5.486h4.467l-2.234-8.046-2.233 8.046z"} },
+  ],[
+    'path',
+    { attrs: {"d":"M39.79 47.617V62h4.084V47.617h4.896V43.89H34.918v3.727h4.873zm22.403-3.727h4.54L57.929 62h-4.568l3.3-6.832-6.295-11.278h4.771l3.78 7.143 3.276-7.143zM16.685 10.762c-1.107 0-2.011.922-2.011 2.051v1.473h6.035v-1.473c0-1.13-.904-2.051-2.012-2.051h-2.012z"} },
+  ],[
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M34.041 18.39c0 9.021-7.172 16.335-16.02 16.335C9.173 34.725 2 27.411 2 18.389 2 9.367 9.173 2.053 18.02 2.053c8.849 0 16.021 7.314 16.021 16.336zM22.09 14.285h2.713c.747 0 1.38.569 1.472 1.322a.75.75 0 01-.014.25l-2.284 9.87c.009.381-.125.753-.374 1.04a1.461 1.461 0 01-1.107.508h-9.533a.698.698 0 01-.692-.705c0-.39.309-.706.692-.706h9.533c.032 0 .055-.01.073-.033a.088.088 0 00.023-.075.752.752 0 01.014-.25l2.266-9.782a.09.09 0 00-.065-.028h-14.23a.698.698 0 01-.692-.705c0-.391.31-.706.692-.706h2.713v-1.473c0-1.91 1.523-3.463 3.396-3.463h2.008c1.873 0 3.395 1.553 3.395 3.463v1.473zm-8.633 8.902c.3 0 .544.25.544.555a.55.55 0 01-.544.555H10.35a.55.55 0 01-.544-.555.55.55 0 01.544-.555h3.106zm-3.226-4.625h3.105c.3 0 .545-.25.545-.555a.551.551 0 00-.545-.556H10.23a.55.55 0 00-.544.555.55.55 0 00.544.556zm4.397 1.792c.3 0 .545.25.545.555 0 .31-.245.56-.55.555H7.726a.55.55 0 01-.545-.555.55.55 0 01.545-.555h6.902z"} },
   ]
 ])
 
@@ -10022,6 +10113,13 @@ export const MIconSafePayment264 = GenerateIcon('4rem', '0 0 64 64', [
   ],[
     'path',
     { attrs: {"d":"M49 27h-3v-4a14 14 0 00-28 0v4h-3a2 2 0 00-2 2v24a2 2 0 002 2h34a2 2 0 002-2V29a2 2 0 00-2-2zm-28-4a11 11 0 0122 0v4H21zm27 29H16V30h32z"} },
+  ]
+])
+
+export const MIconTransfer64 = GenerateIcon('4rem', '0 0 64 64', [
+  [
+    'path',
+    { attrs: {"fill-rule":"evenodd","d":"M45.062 6.936a1.5 1.5 0 10-2.121 2.122l9.44 9.44H13.334a1.5 1.5 0 100 3H52.38l-9.44 9.438a1.5 1.5 0 102.122 2.122l12-12a1.5 1.5 0 000-2.122l-12-12zM21.064 33.058a1.5 1.5 0 10-2.122-2.122l-12 12a1.5 1.5 0 000 2.122l12 12a1.5 1.5 0 102.122-2.122l-9.44-9.439H50.67a1.5 1.5 0 000-3H11.624l9.44-9.44z"} },
   ]
 ])
 
